backend/requirements.txt
fastapi
uvicorn
python-dotenv
notion-client
google-genai
langgraph
langchain
requests
langchain-google-genai
uvicorn[standard]
mistune